Adequate import in cypress commands, move to new helpers module
