+++ packages/volto/cypress/support/helpers.js
@@ -1,0 +1,1 @@
+export * from './volto-slate';
