+++ packages/volto/cypress/tests/core/volto-slate/01-block-slate-title-description.js
@@ -1,4 +1,4 @@
-import { slateBeforeEach } from '../../../support/commands';
+import { slateBeforeEach } from '../../../support/helpers';
 
 describe('Block Tests', () => {
   beforeEach(slateBeforeEach);
