+++ packages/volto/cypress/tests/core/volto-slate/03-block-slate.js
@@ -1,4 +1,4 @@
-import { slateBeforeEach, getSlateBlockValue } from '../../../support/commands';
+import { slateBeforeEach, getSlateBlockValue } from '../../../support/helpers';
 
 describe('Block Tests', () => {
   beforeEach(slateBeforeEach);
