+++ packages/volto/cypress/tests/core/volto-slate/04-block-focus.js
@@ -1,4 +1,4 @@
-import { slateBeforeEach } from '../../../support/commands';
+import { slateBeforeEach } from '../../../support/helpers';
 
 describe('Block Tests', () => {
   beforeEach(slateBeforeEach);
