+++ packages/volto/cypress/tests/core/volto-slate/05-block-slate-format-basics.js
@@ -1,4 +1,4 @@
-import { slateBeforeEach } from '../../../support/commands';
+import { slateBeforeEach } from '../../../support/helpers';
 
 describe('Block Tests: Basic text format', () => {
   beforeEach(slateBeforeEach);
