+++ packages/volto/cypress/tests/core/volto-slate/06-block-slate-format-link.js
@@ -1,4 +1,4 @@
-import { slateBeforeEach } from '../../../support/commands';
+import { slateBeforeEach } from '../../../support/helpers';
 
 describe('Block Tests: Links', () => {
   beforeEach(slateBeforeEach);
