+++ packages/volto/cypress/tests/core/volto-slate/07-block-slate-format-ulist.js
@@ -1,4 +1,4 @@
-import { slateBeforeEach } from '../../../support/commands';
+import { slateBeforeEach } from '../../../support/helpers';
 
 describe('Block Tests: Bulleted lists', () => {
   beforeEach(slateBeforeEach);
