+++ packages/volto/cypress/tests/core/volto-slate/08-block-slate-format-olist.js
@@ -1,4 +1,4 @@
-import { slateBeforeEach } from '../../../support/commands';
+import { slateBeforeEach } from '../../../support/helpers';
 
 describe('Block Tests: Numbered lists', () => {
   beforeEach(slateBeforeEach);
