+++ packages/volto/cypress/tests/core/volto-slate/09-block-slate-slashmenu.js
@@ -1,4 +1,4 @@
-import { slateBeforeEach } from '../../../support/commands';
+import { slateBeforeEach } from '../../../support/helpers';
 
 describe('SlashMenu Test: Shortcuts', () => {
   beforeEach(slateBeforeEach);
