+++ packages/volto/cypress/tests/core/volto-slate/24-block-slate-format-boldlists.js
@@ -1,4 +1,4 @@
-import { slateBeforeEach } from '../../../support/commands';
+import { slateBeforeEach } from '../../../support/helpers';
 
 describe('Block Tests: Bold Bulleted lists', () => {
   beforeEach(slateBeforeEach);
