+++ packages/volto/cypress/tests/core/volto-slate/25-block-slate-paste-images.js
@@ -1,4 +1,4 @@
-import { slateBeforeEach } from '../../../support/commands';
+import { slateBeforeEach } from '../../../support/helpers';
 
 describe('Block Tests: paste external images', () => {
   beforeEach(slateBeforeEach);
