+++ packages/volto/cypress/tests/core/volto-slate/28-table-block-slate-paste.js
@@ -1,4 +1,4 @@
-import { slateBeforeEach } from '../../../support/commands';
+import { slateBeforeEach } from '../../../support/helpers';
 
 describe('Block Tests: pasting content in table block', () => {
   beforeEach(slateBeforeEach);
